# Tiny optimization in state refs lock/unlock