--- a/beacon-chain/state/BUILD.bazel
+++ b/beacon-chain/state/BUILD.bazel
@@ -41,6 +41,7 @@ go_library(
     ],
 )
 
+# gazelle:exclude types_bench_test.go
 go_test(
     name = "go_default_test",
     srcs = [