**What type of PR is this?**

Other/Minor Optimization

**What does this PR do? Why is it needed?**
- The `defer` after release of Go1.4 is almost as good as inlined function call, yet it still has a bit of an overhead.
- I've noticed that `AddRef()` avoids calling `defer`, this PR makes sure that `MinusRef()` also calls unlocking inline.

**Which issues(s) does this PR fix?**

N/A

**Other notes for review**
- To help gauging utility of the PR, here is the output of `benchstats` (~0.5% improvement on speed -- really tiny optimization):
```
± benchstat bench.old bench.new
name                  old time/op    new time/op    delta
Reference_MinusRef-8    30.4ns ± 0%    28.8ns ± 0%   ~     (p=1.000 n=1+1)
```